add fix C4146
see https://docs.microsoft.com/en-us/cpp/error-messages/compiler-warnings/compiler-warning-level-2-c4146?f1url=https%3A%2F%2Fmsdn.microsoft.com%2Fquery%2Fdev15.query%3FappId%3DDev15IDEF1%26l%3DEN-US%26k%3Dk(C4146)%26rd%3Dtrue&view=vs-2019

+++ src/CoinBronKerbosch.cpp
@@ -167,7 +167,7 @@ double CoinBronKerbosch::weightP(const size_t depth, size_t &u) {
         size_t value = P_[depth][t];
 
         while (value) {
-            const size_t node = ((size_t)log2(value & -value)) + (INT_SIZE * t);
+            const size_t node = ((size_t)log2((int)value & -(int)value)) + (INT_SIZE * t);
             weight += vertices_[node].weight;
             value &= value - 1;
 
@@ -233,7 +233,7 @@ void CoinBronKerbosch::bronKerbosch(size_t depth) {
             size_t value = L_[depth][t];
 
             while (value) {
-                const size_t v = ((size_t)log2(value & -value)) + (INT_SIZE * t);
+                const size_t v = ((size_t)log2((int)value & -(int)value)) + (INT_SIZE * t);
                 value &= value - 1;
 
                 //new P and S
